requirements.txt
Flask~=2.3.0
Flask-SQLAlchemy~=3.0.0
python-binance~=1.0.19
Flask-SocketIO~=5.3.6
eventlet~=0.35.2
gunicorn~=21.2.0
flasgger>=0.9.7
